% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_trs.r
\name{bar_trs}
\alias{bar_trs}
\title{Make graphic bar of calculate proportion of type infected containers}
\usage{
bar_trs(df, var, file_name = "./data/g_bar.jpg")
}
\arguments{
\item{`df`}{the dataframe with information}

\item{`var`}{The variable used to calculate proportion of type infected
containers and make graphic bar.}

\item{`path_out`}{path for that graphic bar. By default, it is
"./data/g_bar.jpg"}
}
\value{
make bar graphric with percentages of the typology infect container by
select  variable
}
\description{
Make graphic bar of calculate proportion of type infected containers
}
\examples{
bar_trs(df, "Tipo_de_Estudio", "typeofstudy.jpg")

}
