% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.r
\name{clean_raw_data}
\alias{clean_raw_data}
\title{clean raw data from a entomology study}
\usage{
clean_raw_data(
  df,
  colt = list(Tipo_de_Estudio = col_factor(c("Encuesta", "Verificacion")),
    Clave_Jurisdiccion = "f", Jurisdiccion = "f", Clave_Localidad = "f", Localidad = "f",
    Sector = "f", Fecha_de_Inicio = col_date(format = "\%d/\%m/\%Y"),
    Semana_Epidemiologica = "f", Casas_Revisadas = "d", Casas_Positivas = "d",
    Total_de_Recipientes_con_Agua = "d", Total_de_Recipientes_Positivos = "d"),
  path_out = "~/CursoQR/Package1/rStegomyiaIndices/rStegomyiaIndices/data-raw/qr.csv"
)
}
\arguments{
\item{`df`}{a data.frame with raw data path}

\item{`path_out`}{path for that cleaned data}

\item{`col_name`}{select variables}
}
\value{
Data.frame A .csv file with selected and cleaned data
}
\description{
data.frame chance type data of variable. previously  load with function
"Load_raw_data" file .txt  of entomology study from platform
"Vigilancia Entomológica y Control Integral del Vector", and make a file
.csv of data.frame

chose the variables
that use to calculate entomological risk indices. Fix variable labels
Eliminate spaces in the column names and replace them with underscores _.
A short description...mensionar la plataforma
data.fram from load_raw_data
mensionar todos los pasos
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

df <- load_raw_data(df)

}
