% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maps_stegomyia_indices.R
\name{get_maps_stegomyia_indices}
\alias{get_maps_stegomyia_indices}
\title{get_maps_stegomyia_indices}
\usage{
get_maps_stegomyia_indices(df, m1 = w1, m0 = d0)
}
\arguments{
\item{df}{the dataframe with information.}

\item{m1}{is a shapefile"ejercicio_sectores_hermosillo.shp" with geographic
information of hermosillo city for join dataframe with information of
indices stegomyia}

\item{m0}{is a shapefile "ejercicio_sectores_hermosillo2.shp" with limit of
sectors of hermosillo city. This shapefile use to fution with m1 and give
limits of the sectors and the hermosillo city

  print(maps$HI)  # Print the map for the House Index
  print(maps$CI)  # Print the map for the Container Index
  print(maps$BI)  # Print the map for the Breteau Index}
}
\value{
listmaps: A list of three ggplot objects representing the maps for
different Stegomyia indices.
}
\description{
This function generates maps for three Stegomyia indices (Entomological Risk
Indices) of dataframe make to funtion
get_stegomyia_indices_by_type_of_study_and_geo_is.
This data frame containing information about sectors and their corresponding
indices.
}
\examples{

get_maps_stegomyia_indices(dfm)

}
