% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_and_geo.R
\name{get_stegomyia_indices_by_type_of_study_and_geo}
\alias{get_stegomyia_indices_by_type_of_study_and_geo}
\title{Get stegomyia indices by type of study and geo}
\usage{
get_stegomyia_indices_by_type_of_study_and_geo(df, st = "Verificacion", var)
}
\description{
@description
 The Stegomyia indices are calculated for each sampling select study type
 and geographic variable using the following formulas:
}
\details{
* `Container Index(CI)`: number of infected containers × 100/total number
    of containers
 * `House Index (HI)`: number of infected houses × 100/total number of
    houses
 * `Breteau Index (BI)`: number of positive containers/number of houses
    explored × 100
 @param `df` the dataframe with information
 @param `st` The type of study selected. By default, it is set to
   "Verificacion"
 @param `var` The geographic variable used to calculate the stegomyia
   indices.
 @return
 The dataframe with stegomy indices of the selection of type of study and
   geographic variable.
}
