% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_and_star_date.r
\name{get_stegomyia_indices_by_type_of_study_and_star_date}
\alias{get_stegomyia_indices_by_type_of_study_and_star_date}
\title{Get stegomyia indices by type of study and star date}
\usage{
get_stegomyia_indices_by_type_of_study_and_star_date(
  df,
  st = "Verificacion",
  date = "2021/01/07"
)
}
\arguments{
\item{df}{the dataframe with information}

\item{st}{The type of study selected. By default, it is set to
"Verificacion"}

\item{date}{The date used to calculate the stegomyia indices. By default,
it is set to "2021/01/07".}
}
\value{
The dataframe with Stegomyia indices of the selection of type of study and
date.
}
\description{
The Stegomyia indices are calculated for each sampling. Once the data have
been loaded with function "load_rwa_data" and changed type data of function
"clean_raw_data", select study type and date variable using the following
 formulas:
      - Container Index(CI): (number of infected containers/total number of
      containers) * 100.
      - House Index (HI): (number of infected houses /total number of
      houses) * 100.
      - Breteau Index (BI): (number of positive containers/number of houses
      explored) * 100.
}
\examples{

df<-get_stegomyia_indices_by_type_of_study_and_star_date(df,
st = "Verificacion", date = "2021/01/06" )
}
