% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_star_date_and_geo.R
\name{get_stegomyia_indices_by_type_of_study_star_date_and_geo}
\alias{get_stegomyia_indices_by_type_of_study_star_date_and_geo}
\title{Get stegomyia indices by type of study, date and geo}
\usage{
get_stegomyia_indices_by_type_of_study_star_date_and_geo(
  df,
  st = "Verificacion",
  date = "2021/01/07",
  var
)
}
\description{
@description
 The Stegomyia indices are calculated for each sampling. Once the data have
 been loaded with function "load_rwa_data" and changed type data of function
 "clean_raw_data", select study type, date and geographic variable using the
 following formulas:
}
\details{
* `Container Index(CI)`: (number of infected containers /total number
    of containers) * 100
 * `House Index (HI)`: (number of infected houses /total number of
    houses) * 100
 * `Breteau Index (BI)`: (number of positive containers/number of houses
    explored) * 100
 @param `df` the dataframe with information
 @param `st` The type of study selected. By default, it is set to
   "Verificacion"
 @param `var` The geographic variable used to calculate the stegomyia
   indices
 @param `date` The date used to calculate the stegomyia indices. By default,
  it is set to "2021/01/07".
 @return
 The dataframe with stegomyia indices of the selection of type of study and
   geographic variable.
 @examples
 get_stegomyia_indices_by_type_of_study__star_date_and_geo(df, "Encuesta", "2021/01/07", "Sector")
 @export
}
