% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.r
\docType{data}
\name{valt}
\alias{valt}
\title{clean raw data from a entomology study}
\format{
An object of class \code{list} of length 12.
}
\usage{
valt
}
\arguments{
\item{`df`}{a data.frame with raw data path}

\item{`path_out`}{path for that cleaned data}

\item{`col_name`}{select variables}
}
\value{
Data.frame A .csv file with selected and cleaned data
}
\description{
The function creates a data.frame containing data and chance type data. The
data is assumed to have been previously loaded using the function
"Load_raw_data" from a .txt file associated with an entomology study
conducted on the platform "Vigilancia Entomológica y Control Integral del
Vector". This function aims to generate new data.frame suitable for
further analysis and manipulation. I will also save the data from the
data.frame to a .csv file
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

df <- load_raw_data(df)

}
\keyword{datasets}
