% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.r
\name{clean_raw_data}
\alias{clean_raw_data}
\title{Load and clean raw data from a entomology study}
\usage{
clean_raw_data(
  path,
  path_out = "./data/qr.csv",
  col_name = c("Tipo de Estudio", "Jurisdiccion", "Localidad", "Sector",
    "Fecha de Inicio", "Semana Epidemiologica", "Casas Revisadas", "Casas Positivas",
    "Total de Recipientes con Agua", "Total de Recipientes Positivos")
)
}
\arguments{
\item{path}{a string with raw data path}

\item{path_out}{path for that cleaned data}

\item{col_name}{select variables}
}
\value{
A .csv file with selected and cleaned data
}
\description{
Load and clean raw data from a entomology study
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

col_select1 <- c("Tipo de Estudio",
                 "Jurisdiccion", "Localidad",
                 "Sector", "Fecha de Inicio",
                 "Semana Epidemiologica",
                 "Casas Revisadas",
                 "Casas Positivas",
                 "Total de Recipientes con Agua",
                 "Total de Recipientes Positivos")

df <- load_raw_data(path, col_name = col_select1)

}
