% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_and_star_date.r
\name{get_stegomyia_indices_by_type_of_study_and_star_date}
\alias{get_stegomyia_indices_by_type_of_study_and_star_date}
\title{Get stegomyia indices by type of study and star date
 @description
 The Stegomyia indices are calculated for each sampling select study type
 and date using the following formulas:
 * `Container Index(CI)`: number of infected containers × 100/total number
    of containers
 * `House Index (HI)`: number of infected houses × 100/total number of
    houses
 * `Breteau Index (BI)`: number of positive containers/number of houses
    explored × 100
 @param `df` the dataframe with information
 @param `st` The type of study selected. By default, it is set to
   "Verificacion"
 @param `date` The date used to calculate the stegomyia indices. By default,
  it is set to "2021/01/07".
 @return
 The dataframe with Stegomyia indices of the selection of type of study and
   date.}
\usage{
get_stegomyia_indices_by_type_of_study_and_star_date(
  df,
  st = "Verificacion",
  date = "2021/01/07"
)
}
\description{
Get stegomyia indices by type of study and star date
 @description
 The Stegomyia indices are calculated for each sampling select study type
 and date using the following formulas:
 * `Container Index(CI)`: number of infected containers × 100/total number
    of containers
 * `House Index (HI)`: number of infected houses × 100/total number of
    houses
 * `Breteau Index (BI)`: number of positive containers/number of houses
    explored × 100
 @param `df` the dataframe with information
 @param `st` The type of study selected. By default, it is set to
   "Verificacion"
 @param `date` The date used to calculate the stegomyia indices. By default,
  it is set to "2021/01/07".
 @return
 The dataframe with Stegomyia indices of the selection of type of study and
   date.
}
\examples{
 df<-get_stegomyia_indices_by_type_of_study_and_star_date(df, st = "Verificacion",
 date = "2021/01/06" )

}
