% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_typology_container_by_type_of_study_and_location.R
\name{get_typology_container_by_type_of_study_and_location}
\alias{get_typology_container_by_type_of_study_and_location}
\title{Positive Typology Container by type of study and location}
\usage{
get_typology_container_by_type_of_study_and_location(
  df,
  st = "Verificacion",
  var = c("CIUDAD_OBREGÓN", "HERMOSILLO")
)
}
\arguments{
\item{df}{the dataframe with information.}

\item{st}{The type of study selected. By default, it is set to
"Verificacion".}

\item{var}{The location with geographic variable used to .}
}
\value{
The dataframe with ######### of the selection of type of study and
geographic variable.
}
\description{
Positive Typology Container by type of study and location
}
\examples{
get_typology_container_by_type_of_study_and_location(df, st = "Verificacion",
  var ="CIUDAD_OBREGÓN").

}
