% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.r
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw data  of entomology study}
\usage{
load_raw_data(
  path,
  col_name = c("Tipo de Estudio", "Jurisdiccion", "Localidad", "Sector",
    "Fecha de Inicio", "Semana Epidemiologica", "Casas Revisadas", "Casas Positivas",
    "Total de Recipientes con Agua", "Total de Recipientes Positivos")
)
}
\arguments{
\item{`path`}{a string with raw data path}

\item{`col_name`}{names of variables,
,}
}
\value{
A data frame with selected and format in variable data
}
\description{
Load raw data from a .txt file of entomology study from platform
"Vigilancia Entomológica y Control Integral del Vector",chose the variables
that use to calculate entomological risk indices. Fix variable labels
Eliminate spaces in the column names and replace them with underscores _. If
one of the locality and/or jurisdiction variables was chosen, new columns
will be created separating the number of the variables and their name
Load raw data from a .txt file of entomology study from nombre de pla plataforma

A short description...
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

col_select <- c("Tipo de Estudio",
                 "Jurisdiccion", "Localidad",
                 "Sector", "Fecha de Inicio",
                 "Semana Epidemiologica",
                 "Casas Revisadas",
                 "Casas Positivas",
                 "Total de Recipientes con Agua",
                 "Total de Recipientes Positivos")

df <- load_raw_data(path, col_name = col_select)

}
