% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.r
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw data from a entomology study}
\usage{
load_raw_data(
  path,
  col_name = c("Tipo de Estudio", "Jurisdiccion", "Localidad", "Sector",
    "Fecha de Inicio", "Semana Epidemiologica", "Casas Revisadas", "Casas Positivas",
    "Total de Recipientes con Agua", "Total de Recipientes Positivos")
)
}
\arguments{
\item{`path`}{a string with raw data path}

\item{`col_name`}{select variables}
}
\value{
A data frame with selected and format in variable data
}
\description{
Load raw data from a entomology study
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

col_select <- c("Tipo de Estudio",
                 "Jurisdiccion", "Localidad",
                 "Sector", "Fecha de Inicio",
                 "Semana Epidemiologica",
                 "Casas Revisadas",
                 "Casas Positivas",
                 "Total de Recipientes con Agua",
                 "Total de Recipientes Positivos")

df <- load_raw_data(path, col_name = col_select)

}
