% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw data  of entomology study}
\usage{
load_raw_data(
  path,
  col_name = c("Tipo de Estudio", "Jurisdiccion", "Localidad", "Sector",
    "Fecha de Inicio", "Semana Epidemiologica", "Casas Revisadas", "Casas Positivas",
    "Total de Recipientes con Agua", "Total de Recipientes Positivos")
)
}
\arguments{
\item{`path`}{a string with raw data path}

\item{`col_name`}{names of variables,
,}
}
\value{
A data frame with selected and format in variable data
}
\description{
Load_raw_data from a .txt file of the entomology study conducted on the
platform 'Vigilancia Entomológica y Control Integral del Vector.' Choose the
variables that will be used to calculate entomological risk indices. Fix
variable labels and eliminate spaces in the column names, replacing them
with underscores (_). If one of the "Localidad", "Municipio"and/or
"Jurisdiccion" variables was selected, new columns will be created,
separating the variable of number  and other with name.
}
\examples{

path <- "./data-raw/estudio_entomologico.txt"

col_select <- c("Tipo de Estudio",
                 "Jurisdiccion", "Localidad",
                 "Sector", "Fecha de Inicio",
                 "Semana Epidemiologica",
                 "Casas Revisadas",
                 "Casas Positivas",
                 "Total de Recipientes con Agua",
                 "Total de Recipientes Positivos")

df <- load_raw_data(path, col_name = col_select)

}
